scripts/requirements.txt
faker
python-dotenv
supabase
